Modernize CMake
Change Log:
* CMake dependency bumped to 3.10.
* Moved to target-oriented CMake set-up.
* ExSeisDat can be included in a project via add_subdirectory.
* Compiled/installed ExSeisDat can be used with find_package from the
  ExSeisDat build directory or install directory.
* FFTW3 and MPI now linked as imported targets.
* Added options to enable/disable specific tests
* Added CMake option to add development flags used by the CI environment
  for GCC and Clang.

+++ test/cwraptests/main.c
@@ -353,7 +353,7 @@ int main()
     piol_file_write_interface_write_text(
         write_direct, "Test WriteInterface Text");
 
-    const int write_direct_ns = 700;
+    const size_t write_direct_ns = 700;
     piol_file_write_interface_write_ns(write_direct, write_direct_ns);
 
     piol_file_write_interface_write_nt(write_direct, 710);
